requirements.txt
crewai
langchain-openai
python-dotenv
fastapi
uvicorn
pydantic
jinja2
python-multipart 